feature(im):即时通讯基本功能实现

+++ tt-zhipin-server/tt-zhipin-im/tt-zhipin-im-core/src/main/java/com/whoiszxl/zhipin/im/pack/AddFriendPack.java
@@ -18,7 +18,7 @@ public class AddFriendPack extends Packet {
 
 
     @Override
-    public Byte getCommand() {
+    public Integer getCommand() {
         return Command.ADD_FRIEND;
     }
 }
