+++ tt-zhipin-server/tt-zhipin-im/tt-zhipin-im-core/src/main/java/com/whoiszxl/zhipin/im/pack/LogoutPack.java
@@ -17,7 +17,7 @@ public class LogoutPack extends Packet {
     private String memberId;
 
     @Override
-    public Byte getCommand() {
+    public Integer getCommand() {
         return Command.LOGOUT;
     }
 }
